feat: implement embedding logic with SVD

+++ src/config.rs
@@ -3,11 +3,11 @@ use derive_builder::Builder;
 #[derive(Debug, Clone, Builder)]
 pub struct WatermarkConfig {
     /// 嵌入强度 (对应 d1)
-    #[builder(default = "36.0")]
-    pub strength_1: f32,
+    #[builder(default = "36")]
+    pub strength_1: i32,
     /// 嵌入强度 (对应 d2)
-    #[builder(default = "20.0")]
-    pub strength_2: f32,
+    #[builder(default = "20")]
+    pub strength_2: i32,
     /// 随机种子/密码
     #[builder(default = "WatermarkMode::Normal")]
     pub mode: WatermarkMode,
@@ -22,8 +22,8 @@ pub enum WatermarkMode {
 impl Default for WatermarkConfig {
     fn default() -> Self {
         Self {
-            strength_1: 36.0,
-            strength_2: 20.0,
+            strength_1: 36,
+            strength_2: 20,
             mode: WatermarkMode::Normal,
         }
     }
